Fixed Rune creation when mining obsidian.
The Rune Shell logic was treating any 5x5 block of obsidian hit in the center as a rune. This was causing problems when mining obsidian, as the player would often accidentally create runes, which would promptly burst into flames when hit again.

I modified the rune shell logic to disallow obsidian in the signature blocks. This limits the possibilities for the number of unique signatures you can create, but it's a very small trade-off considering the alternative.