Added Op command to list registered runes.
Previously, there was no way to get a list of the runes that were registered. This caused problems when attempting to investigate bugs, or just administer the server in general.

I added the /blinklist command, which can only be executed by an Op. The command will list all Blink runes registered on the server. Each rune will be listed with the name of the world it is located in, it's x, y and z coordinates, and the signature group that it belongs to. The runes are also listed in the same order that teleportation would occur.