Convert save file to JSON
__Problem__
The current method of saving runes serializes a location object. This tightly couples the save file to a specific class, and makes it difficult to refactor the project.

__Resolution__
The loading mechanism now attempts load a legacy file if it exists, and if not then attempts to load a json file instead.
The saving mechanism will only save a json file. If a legacy file is detected after a json file is saved, it will be deleted.