Fix assumption of saves directory
Problem:
The mod is assuming that all worlds are saved in the "saves" directories. This is true when running locally, but is not true when running in a dedicated server environment.

Fix:
I modified the logic to check for the existence of a "saves" directory. If it exists, the mod will attempt to save/load rune data from world directories within the saves directory. If the "saves" directory does not exist, the mod will assume that the world directories are stored within the root of the working directory.