Fix servers don't load runes
Problem:
Runes are not being loaded when using the mod on a dedicated server.

Cause:
The WorldEvent.Load event is not being fired on dedicated servers. This event is what's being monitored to determine when to load the runes.

Fix:
I changed the loading of the runes to be triggered by the FMLServerStartedEvent instead.