Fix Blink instance is not injected
Problem:
Some versions of forge are not injecting the Blink instance.

Cause:
It seems that the instance annotation requires the mod id and not the mod name. The mod id is lowercase, but the name has an uppercase first character.

Fix:
I modified the annotation to use the lowercase mod id.